Move criteriasource useEffect to currentPathway
