+++ src/components/CurrentPathwayProvider.tsx
@@ -12,6 +12,8 @@ import { Pathway } from 'pathways-model';
 import useRefUndoState from 'hooks/useRefUndoState';
 import { usePathwaysContext } from './PathwaysProvider';
 import HotKeys from 'react-hot-keys';
+import { useCriteriaContext } from './CriteriaProvider';
+import produce from 'immer';
 
 interface CurrentPathwayContextInterface {
   pathway: Pathway | null;
@@ -43,6 +45,7 @@ export const CurrentPathwayProvider: FC<CurrentPathwayProviderProps> = memo(({ c
     _resetPathway,
     _setPathway
   ] = useRefUndoState<Pathway | null>(null);
+  const { criteria } = useCriteriaContext();
   const { updatePathway } = usePathwaysContext();
 
   const undoPathway = useCallback(() => {
@@ -70,6 +73,33 @@ export const CurrentPathwayProvider: FC<CurrentPathwayProviderProps> = memo(({ c
   useEffect(() => {
     if (pathway) updatePathway(pathway);
   }, [pathway, updatePathway]);
+
+  // Update criteriaSource if criteria change
+  useEffect(() => {
+    if (!pathway) return;
+
+    const criteriaIds = criteria.map(crit => crit.id);
+    const newPathway = produce(pathway, draftPathway => {
+      Object.entries(draftPathway.nodes).forEach(([nodeIndex, node]) => {
+        node.transitions.forEach(({ condition }, transitionIndex) => {
+          // If a matching criteria does not already exist, try and find one
+          if (condition && !criteriaIds.includes(condition.criteriaSource as string)) {
+            const [library, statement] = condition.cql.split('.');
+            const criteriaSource = criteria.find(
+              crit => crit.elm?.library.identifier.id === library && crit.statement === statement
+            )?.id;
+            // Only update if a criteria source is actually found.
+            if (criteriaSource) {
+              const condition =
+                draftPathway.nodes[nodeIndex].transitions[transitionIndex].condition;
+              if (condition) condition.criteriaSource = criteriaSource;
+            }
+          }
+        });
+      });
+    });
+    resetCurrentPathway(newPathway);
+  }, [criteria, pathway, resetCurrentPathway]);
 
   return (
     <CurrentPathwayContext.Provider
